perf(types): migrated to dts, typed components better, exported types BREAKING CHANGE: might require your input if you you choose to upgrade
