--- .travis.yml
@@ -1,27 +1,0 @@
-sudo: false
-
-language: node_js
-node_js:
-  - 12
-
-install:
-  - yarn
-
-cache:
-  directories:
-    - ~/.npm
-notifications:
-  email: false
-
-before_install:
-  - yarn && yarn add codecov.io coveralls
-
-after_success:
-  - cat ./coverage/lcov.info | ./node_modules/codecov.io/bin/codecov.io.js
-  - cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js
-  - yarn run build
-  - yarn run travis-deploy-once "yarn run semantic-release"
-
-branches:
-  only:
-    - master
