+++ example/vite.config.js
@@ -1,0 +1,11 @@
+import { defineConfig } from 'vite';
+import ReactPlugin from 'vite-preset-react';
+
+// https://vitejs.dev/config/
+export default defineConfig({
+  plugins: [
+    ReactPlugin({
+      injectReact: false,
+    }),
+  ],
+});
