+++ package.json
@@ -1,6 +1,9 @@
 {
+  "name": "react-flex-ready",
   "version": "1.2.7",
   "license": "MIT",
+  "author": "smakosh",
+  "module": "dist/react-flex-ready.esm.js",
   "main": "dist/index.js",
   "typings": "dist/index.d.ts",
   "files": [
@@ -8,25 +11,20 @@
     "src"
   ],
   "engines": {
-    "node": ">=10"
+    "node": ">=12"
   },
   "scripts": {
-    "start": "tsdx watch",
-    "build": "tsdx build",
-    "test": "tsdx test --passWithNoTests",
-    "lint": "tsdx lint",
-    "prepare": "tsdx build",
+    "start": "dts watch",
+    "build": "dts build",
+    "test": "dts test --passWithNoTests",
+    "lint": "dts lint",
+    "prepare": "dts build",
     "size": "size-limit",
-    "analyze": "size-limit --why",
-    "travis-deploy-once": "travis-deploy-once",
-    "semantic-release": "semantic-release"
-  },
-  "peerDependencies": {
-    "react": ">=16"
+    "analyze": "size-limit --why"
   },
   "husky": {
     "hooks": {
-      "pre-commit": "tsdx lint"
+      "pre-commit": "dts lint"
     }
   },
   "prettier": {
@@ -35,9 +33,6 @@
     "singleQuote": true,
     "trailingComma": "es5"
   },
-  "name": "react-flex-ready",
-  "author": "smakosh",
-  "module": "dist/react-flex-ready.esm.js",
   "size-limit": [
     {
       "path": "dist/react-flex-ready.cjs.production.min.js",
@@ -48,25 +43,31 @@
       "limit": "16 KB"
     }
   ],
+  "peerDependencies": {
+    "react": ">=16"
+  },
+  "dependencies": {
+    "styled-components": "^5.3.3"
+  },
   "devDependencies": {
     "@babel/core": "^7.11.6",
-    "@size-limit/preset-small-lib": "^5.0.1",
+    "@size-limit/preset-small-lib": "^7.0.8",
+    "@tsconfig/create-react-app": "^1.0.2",
+    "@tsconfig/recommended": "^1.0.1",
     "@types/react": "^16.9.50",
     "@types/react-dom": "^16.9.8",
     "@types/styled-components": "^5.1.3",
     "babel-loader": "^8.1.0",
+    "dts-cli": "^1.4.0",
     "husky": "^4.3.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "react-is": "^16.13.1",
-    "semantic-release": "^17.4.4",
-    "size-limit": "^5.0.1",
-    "travis-deploy-once": "^5.0.11",
-    "tsdx": "^0.14.0",
+    "size-limit": "^7.0.8",
     "tslib": "^2.0.1",
-    "typescript": "^4.0.3"
+    "typescript": "^4.6.2"
   },
-  "dependencies": {
-    "styled-components": "^5.2.0"
+  "jest": {
+    "testEnvironment": "jsdom"
   }
 }
