+++ src/index.ts
@@ -1,2 +1,3 @@
 export { default as Flex } from './Flex';
 export { default as Item } from './Item';
+export type { Range, FlexProps, ItemProps } from './interfaces';
