+++ test/item.test.tsx
@@ -5,7 +5,7 @@ import { Item } from '../src';
 describe('Item', () => {
   it('renders without crashing', () => {
     const div = document.createElement('div');
-    ReactDOM.render(<Item />, div);
+    ReactDOM.render(<Item col={2} />, div);
     ReactDOM.unmountComponentAtNode(div);
   });
 });
